src/requirements.txt
requests
python-dotenv
sparksql
boto3